update files as per mvc logic
